refactor out functions into state helpers
