+++ public/app/plugins/datasource/prometheus/querybuilder/components/metric-encyclopedia/LetterSearch.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 
 import { useTheme2 } from '@grafana/ui';
 
-import { alphabet, alphabetCheck } from './helpers';
 import { getStyles } from './styles';
 import { MetricData, MetricsData } from './types';
 
@@ -60,3 +59,65 @@ export function LetterSearch(props: LetterSearchProps) {
     </div>
   );
 }
+
+function alphabetCheck() {
+  const alphabetDict: { [char: string]: number } = {
+    A: 0,
+    B: 0,
+    C: 0,
+    D: 0,
+    E: 0,
+    F: 0,
+    G: 0,
+    H: 0,
+    I: 0,
+    J: 0,
+    K: 0,
+    L: 0,
+    M: 0,
+    N: 0,
+    O: 0,
+    P: 0,
+    Q: 0,
+    R: 0,
+    S: 0,
+    T: 0,
+    U: 0,
+    V: 0,
+    W: 0,
+    X: 0,
+    Y: 0,
+    Z: 0,
+  };
+
+  return alphabetDict;
+}
+
+export const alphabet = [
+  'A',
+  'B',
+  'C',
+  'D',
+  'E',
+  'F',
+  'G',
+  'H',
+  'I',
+  'J',
+  'K',
+  'L',
+  'M',
+  'N',
+  'O',
+  'P',
+  'Q',
+  'R',
+  'S',
+  'T',
+  'U',
+  'V',
+  'W',
+  'X',
+  'Y',
+  'Z',
+];
